requirements.txt
amqp==5.2.0
asgiref==3.8.1
billiard==4.2.0
black==24.4.2
cachetools==5.3.3
celery==5.4.0
certifi==2024.7.4
cffi==1.16.0
cfgv==3.4.0
charset-normalizer==3.3.2
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
colorama==0.4.6
cron-descriptor==1.4.3
cryptography==42.0.8
distlib==0.3.8
dj-database-url==2.2.0
Django==5.0.7
django-cacheops==7.0.2
django-celery-beat==2.6.0
django-cleanup==8.1.0
django-cors-headers==4.4.0
django-debug-toolbar==4.4.6
django-querycount==0.8.3
django-rest-knox==4.2.0
django-storages==1.14.4
django-timezone-field==7.0
djangorestframework==3.15.2
environs==11.0.0
filelock==3.15.4
flower==2.0.1
funcy==2.0
google-api-core==2.19.1
google-auth==2.32.0
google-cloud-core==2.4.1
google-cloud-storage==2.17.0
google-crc32c==1.5.0
google-resumable-media==2.7.1
googleapis-common-protos==1.63.2
gunicorn==22.0.0
humanize==4.10.0
identify==2.6.0
idna==3.7
kombu==5.3.7
marshmallow==3.21.3
mypy==1.10.1
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.1
pathspec==0.12.1
pillow==10.4.0
platformdirs==4.2.2
pre-commit==3.7.1
prometheus_client==0.20.0
prompt_toolkit==3.0.47
proto-plus==1.24.0
protobuf==5.27.2
psycopg2-binary==2.9.9
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
python-crontab==3.2.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
redis==5.0.7
requests==2.32.3
rsa==4.9
sentry-sdk==2.9.0
setuptools==70.3.0
six==1.16.0
sqlparse==0.5.0
tornado==6.4.1
typed_ast==1.5.5
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.2
vine==5.1.0
virtualenv==20.26.3
wcwidth==0.2.13
whitenoise==6.7.0